Fix spending points
Points are now spent correctly. The application used to keep all transactions in one massive list and then handle spending by removing from the end of the list. In an effort to more towards a more object oriented approach, each payer keeps track of its own transactions. The spending function had to be redone so that it knows how to go throw the different payer point balances and remove the correct amount of points from the payer with the oldest transactions.